+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Most recent releases are shown at the top. Each release shows:
 
 ### changed
 - **Breaking Changes**: SparseStore and DenseStore now created using factories  (#191)
+- **Breaking Changes**: Changed `k` parameter to `limit` parameter in `LLM.ask`  (#194)
 - Added `return_generator` argument to `SparseStore` (#187)
 - Added `custom_vectorstore` and `reset` parameters to `LLM.load_vectorstore`. (#192)